Fix highlighting of interpolated strings in .html.erb files
In code like:

    <a href="<%= link_to :whatever %>">

the scope that contains the ERB tags is a ".text .string" scope, not a
".source .string" scope like our styles were expecting.

We now don't care about any scopes outside of the string itself when
detecting string interpolation. If there is source code inside a string,
it is considered to be interpolation, no matter where that string lives.